# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Task Tracker API
  version: 0.1.0

paths:
  /tasks:
    get:
      summary: Get all tasks
      operationId: listTasks
      tags:
        - tasks
      responses:
        '200':
          description: List of all tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

    post:
      summary: Create a new task
      operationId: createTask
      tags:
        - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTask'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  /tasks/{id}:
    get:
      summary: Get a task by ID
      operationId: getTask
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found

  /tasks/{id}/done:
    put:
      summary: Mark a task as done
      operationId: markTaskDone
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task marked as done
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found

  /tasks/{id}/title:
    put:
      summary: Change the title of a task
      operationId: renameTask
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameTask'
      responses:
        '200':
          description: Task renamed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        done:
          type: boolean
      required: [ id, title, done ]

    NewTask:
      type: object
      properties:
        title:
          type: string
      required: [ title ]

    RenameTask:
      type: object
      properties:
        title:
          type: string
      required: [ title ]